Add files via upload
IMPORTANT : This was a group project. I worked with Charilaos Kapelonis , you can check his github at : ckapelonis02.

The project implementation is provided with the project description. Unfortunatelly the project description is provided only in Greek and its too wordy to translate it at the very moment... I will try to work on that later..